+++ examples/simple-app/src/App.js
@@ -2,8 +2,8 @@ import { useDataQuery } from '@dhis2/app-runtime'
 import moment from 'moment'
 import React from 'react'
 import { Alerter } from './Alerter.js'
-import style from './App.style'
-import i18n from './locales'
+import style from './App.style.js'
+import i18n from './locales/index.js'
 
 const query = {
     me: {